# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Subscription Reminder API",
    "version": "1.0.0",
    "description": "API documentation for the Subscription Reminder application"
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Development server"
    },
    {
      "url": "https://subscribe_mind.sixpx.workers.dev",
      "description": "Production server"
    },
    {
      "url": "https://mind.sixmillions.cn",
      "description": "Alternative production server"
    }
  ],
  "paths": {
    "/api/auth/{token}": {
      "get": {
        "summary": "Authenticate user with token",
        "description": "Validate the provided token for authentication",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authentication token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful authentication",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/mind": {
      "get": {
        "summary": "Get all minds",
        "description": "Retrieve all minds (reminders) from the database",
        "responses": {
          "200": {
            "description": "Successful retrieval of minds",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new mind",
        "description": "Create a new mind (reminder) with the provided details",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Title of the mind"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the mind"
                  },
                  "time": {
                    "type": "string",
                    "description": "Scheduled time for the mind"
                  },
                  "trigger": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "List of triggers for the mind"
                  },
                  "enabled": {
                    "type": "boolean",
                    "description": "Whether the mind is enabled or not"
                  }
                },
                "required": ["title", "time"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful creation of mind",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/mind/{id}": {
      "get": {
        "summary": "Get a specific mind by ID",
        "description": "Retrieve a specific mind (reminder) by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the mind to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of mind",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Mind not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a mind by ID",
        "description": "Update an existing mind (reminder) with the provided details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the mind to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Title of the mind"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the mind"
                  },
                  "time": {
                    "type": "string",
                    "description": "Scheduled time for the mind"
                  },
                  "trigger": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "List of triggers for the mind"
                  },
                  "enabled": {
                    "type": "boolean",
                    "description": "Whether the mind is enabled or not"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful update of mind",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Mind not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a mind by ID",
        "description": "Delete a specific mind (reminder) by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the mind to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful deletion of mind",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Mind not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/close/{id}": {
      "get": {
        "summary": "Close a mind by ID",
        "description": "Set the enabled flag of a mind to false",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the mind to close"
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authentication token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful closing of mind",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Mind not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/trigger": {
      "get": {
        "summary": "Get all triggers",
        "description": "Retrieve all triggers from the database",
        "responses": {
          "200": {
            "description": "Successful retrieval of triggers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or update a trigger",
        "description": "Create a new trigger or update an existing trigger with the provided configuration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "description": "Key/name of the trigger"
                  },
                  "config": {
                    "type": "object",
                    "description": "Configuration for the trigger"
                  }
                },
                "required": ["key", "config"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful creation or update of trigger",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/trigger/{key}": {
      "delete": {
        "summary": "Delete a trigger by key",
        "description": "Delete a specific trigger by its key",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Key of the trigger to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful deletion of trigger",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Trigger not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/history": {
      "get": {
        "summary": "Get execution history",
        "description": "Retrieve the execution history of minds",
        "responses": {
          "200": {
            "description": "Successful retrieval of history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "his_list": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "title": {
                                "type": "string"
                              },
                              "executionTime": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add execution history record",
        "description": "Add a new execution history record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "ID of the mind"
                  },
                  "title": {
                    "type": "string",
                    "description": "Title of the mind"
                  },
                  "executionTime": {
                    "type": "string",
                    "description": "Execution time"
                  },
                  "status": {
                    "type": "string",
                    "description": "Status of execution"
                  }
                },
                "required": ["id", "title", "executionTime", "status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful addition of history record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clear execution history",
        "description": "Clear all execution history records",
        "responses": {
          "200": {
            "description": "Successful clearing of history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/mail": {
      "get": {
        "summary": "Send test email",
        "description": "Send a test email (demo endpoint)",
        "responses": {
          "200": {
            "description": "Successful email sending",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ding": {
      "get": {
        "summary": "Send test DingTalk message",
        "description": "Send a test DingTalk message (demo endpoint)",
        "responses": {
          "200": {
            "description": "Successful message sending",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}